+++ accessory/HumidifierDehumidifier.js
@@ -1,31 +1,31 @@
 'use strict';
 
 const {Accessory} = require('../util/Accessory');
-const {addCurrentRelativeHumidityCharacteristic} = require('./characteristic/ClimateThermostat');
 const {addSwingModeCharacteristic, addActiveCharacteristic} = require('./characteristic/BinarySensor');
 const {
     addWaterLevelCharacteristic,
     addRotationSpeedCharacteristic,
-    addRelativeHumidityDehumidifierThresholdCharacteristic,
+    addCurrentRelativeHumidityCharacteristic,
     addRelativeHumidityHumidifierThresholdCharacteristic,
+    addRelativeHumidityDehumidifierThresholdCharacteristic
+} = require('./characteristic/Climate');
+const {
     addCurrentHumidifierDehumidifierStateCharacteristic,
     addTargetHumidifierDehumidifierStateCharacteristic
 } = require('./characteristic/ClimateDeHumidifier');
-
-
 
 class HumidifierDehumidifierAccessory extends Accessory {
     constructor(platform, config) {
         super(platform, config);
 
         this._services = [
-            this._getAccessoryInformationService('Humidifier Dehumidifier'),
+            this._getAccessoryInformationService('Humidifier/Dehumidifier'),
             this._getPrimaryService()
         ]
     }
 
     _getPrimaryService() {
-        this._log.debug(`Creating Humidifier Dehumidifier service for ${this.name}`);
+        this._log.debug(`Creating Humidifier/Dehumidifier service for ${this.name}`);
         let primaryService = new this.Service.HumidifierDehumidifier(this.name);
         addCurrentRelativeHumidityCharacteristic.bind(this)(primaryService);
         addCurrentHumidifierDehumidifierStateCharacteristic.bind(this)(primaryService);